# Proof of concept: shape iv ancestor search